Feature/bm 2

+++ src/Service/PlayerService.php
@@ -2,7 +2,6 @@
 
 namespace App\Service;
 
-use App\Controller\PlayerController;
 use App\Entity\Player;
 use App\Entity\PlayerAttribute;
 use App\Entity\PlayerAward;
@@ -10,6 +9,7 @@ use App\Entity\PlayerStats;
 use App\Entity\Season;
 use App\Entity\Team;
 use App\Entity\TrainingCamp;
+use App\Event\SetPlayerJerseyNumberEvent;
 use App\Factory\Factory\PlayerDetailsFactory;
 use App\Factory\Model\PlayerDetailsModel;
 use App\Model\TeamDraft;
@@ -17,8 +17,8 @@ use App\Repository\PlayerRepository;
 use App\Repository\PlayerStatsRepository;
 use App\Repository\TeamRepository;
 use App\Utils\Award;
-use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 
 /**
  * Class PlayerService
@@ -27,54 +27,29 @@ use Doctrine\ORM\EntityManagerInterface;
  */
 class PlayerService
 {
-    const DRAFT_PICK_SALARY_A_GAME = 351;
+    const DRAFT_PICK_SALARY_A_GAME = 351.0;
     const ADD_TO_SALARY_FOR_A_RING = 25;
 
-    /**
-     * @var EntityManagerInterface
-     */
     private $entityManager;
-
-    /**
-     * @var TeamRepository
-     */
     private $teamRepository;
-
-    /**
-     * @var SeasonService
-     */
     private $seasonService;
-
-    /**
-     * @var ServerService
-     */
     private $serverService;
-
-    /**
-     * @var PlayerStatsRepository
-     */
     private $playerStatsRepository;
-
-    /**
-     * @var PlayerRepository
-     */
     private $playerRepository;
-
-    /**
-     * @var PlayerDetailsFactory
-     */
     private $playerDetailsFactory;
+    private $eventDispatcher;
 
     /**
      * PlayerService constructor.
      *
-     * @param EntityManagerInterface $entityManager
-     * @param TeamRepository $teamRepository
-     * @param SeasonService $seasonService
-     * @param ServerService $serverService
-     * @param PlayerStatsRepository $playerStatsRepository
-     * @param PlayerRepository $playerRepository
-     * @param PlayerDetailsFactory $playerDetailsFactory
+     * @param EntityManagerInterface   $entityManager
+     * @param TeamRepository           $teamRepository
+     * @param SeasonService            $seasonService
+     * @param ServerService            $serverService
+     * @param PlayerStatsRepository    $playerStatsRepository
+     * @param PlayerRepository         $playerRepository
+     * @param PlayerDetailsFactory     $playerDetailsFactory
+     * @param EventDispatcherInterface $eventDispatcher
      */
     public function __construct(
         EntityManagerInterface $entityManager,
@@ -83,43 +58,27 @@ class PlayerService
         ServerService $serverService,
         PlayerStatsRepository $playerStatsRepository,
         PlayerRepository $playerRepository,
-        PlayerDetailsFactory $playerDetailsFactory
+        PlayerDetailsFactory $playerDetailsFactory,
+        EventDispatcherInterface $eventDispatcher
     ) {
-        $this->entityManager = $entityManager;
-        $this->teamRepository = $teamRepository;
-        $this->seasonService = $seasonService;
-        $this->serverService = $serverService;
+        $this->entityManager         = $entityManager;
+        $this->teamRepository        = $teamRepository;
+        $this->seasonService         = $seasonService;
+        $this->serverService         = $serverService;
         $this->playerStatsRepository = $playerStatsRepository;
-        $this->playerRepository = $playerRepository;
-        $this->playerDetailsFactory = $playerDetailsFactory;
+        $this->playerRepository      = $playerRepository;
+        $this->playerDetailsFactory  = $playerDetailsFactory;
+        $this->eventDispatcher       = $eventDispatcher;
     }
 
     /**
-     * @param Team $team
+     * @param Team   $team
      * @param Player $player
-     * @param bool $flush
      */
-    public function setPlayerJerseyNumber(Team $team, Player $player, bool $flush = true)
+    public function setPlayerJerseyNumber(Team $team, Player $player)
     {
-        /**
-         * @var array $usedJerseys
-         */
-        $usedJerseys[] = $this->playerRepository->getTakenJerseyNumbers($team);
-        $jerseyNumber = null;
-
-        while($jerseyNumber == null) {
-            $generatedNumber = rand(0, 99);
-            if(!in_array($generatedNumber, $usedJerseys)) {
-                $jerseyNumber = $generatedNumber;
-            }
-        }
-
-        $player->setJerseyNumber($jerseyNumber);
-
-        if($flush) {
-            $this->entityManager->persist($player);
-            $this->entityManager->flush();
-        }
+        $event = new SetPlayerJerseyNumberEvent($player, $team);
+        $this->eventDispatcher->dispatch($event, SetPlayerJerseyNumberEvent::NAME);
     }
 
     /**
@@ -131,70 +90,58 @@ class PlayerService
      */
     public function draftPlayer(Player $player): array
     {
-        $teamName = '';
-        $drafted = null;
-        $teamNumber = 1;
+        $teamName = null;
+        $draftPick = 0;
 
-        /**
-         * @var Team[] $teams
-         */
         $teams = $this->teamRepository->findBy([
             'server' => $player->getServer()
         ]);
 
-        if($teams) {
-            $teamArray = [];
-
-            foreach($teams as $teamCheck) {
-                if(count($teamCheck->getRealPlayers()) < Team::DEFAULT_PLAYER_LIMIT_IN_TEAM) {
-                    $teamArray[] = new TeamDraft($teamCheck);
-                }
-            }
-
-            if(!empty($teamArray)) {
-                shuffle($teamArray);
-
-                $teamNumber = count($teamArray);
-                $selectedTeamNumber = rand(0, $teamNumber - 1);
-
-                /**
-                 * @var TeamDraft $team
-                 */
-                $team = $teamArray[$selectedTeamNumber];
-
-                self::checkAndFixFakePlayerLimitOnSigning($team->getTeam(), $player);
-
-                $teamName = $team->getTeam()->getCity() . ' ' . $team->getTeam()->getName();
-                $salary = self::DRAFT_PICK_SALARY_A_GAME;
-
-                $player->setTeam($team->getTeam());
-                $player->setContractYears(2);
-                $player->setContractSalary($salary);
-                $player->setSeasonEndsContract(
-                    $this->seasonService->getActiveSeason($player->getServer())->getId() + 2
-                );
-
-                $this->entityManager->persist($player);
-                $this->entityManager->flush();
-
-                $drafted = $team;
-            } else {
-                $draftPick = 0;
-            }
-        } else {
-            $draftPick = 0;
+        if (count($teams) === 0) {
+            return [
+                [$teamName],
+                [$draftPick]
+            ];
         }
 
-        if($drafted) {
-            $draftPick = rand(0, $teamNumber * 2) + 1;
-            self::setPlayerJerseyNumber($player->getTeam(), $player);
-        } else {
-            $draftPick = 0;
+        $teamArray = [];
+
+        foreach ($teams as $teamCheck) {
+            if (count($teamCheck->getRealPlayers()) < Team::DEFAULT_PLAYER_LIMIT_IN_TEAM) {
+                $teamArray[] = new TeamDraft($teamCheck);
+            }
         }
+
+        if (empty($teamArray)) {
+            return [
+                [$teamName],
+                [$draftPick]
+            ];
+        }
+
+        shuffle($teamArray);
+
+        $selectedTeamNumber = rand(0, count($teamArray) - 1);
+        $team = $teamArray[$selectedTeamNumber];
+        $draftPick = rand(0, count($teamArray) * 2) + 1;
+
+        $this->checkAndFixFakePlayerLimitOnSigning($team->getTeam(), $player);
+
+        $player->setTeam($team->getTeam());
+        $player->setContractYears(2);
+        $player->setContractSalary(self::DRAFT_PICK_SALARY_A_GAME);
+        $player->setSeasonEndsContract(
+            $this->seasonService->getActiveSeason($player->getServer())->getId() + 2
+        );
+
+        $this->entityManager->persist($player);
+        $this->entityManager->flush();
+
+        $this->setPlayerJerseyNumber($player->getTeam(), $player);
 
         return [
-            [$teamName],
-            [$draftPick]
+            $team->getTeam()->getFullTeamName(),
+            $draftPick
         ];
     }
 
@@ -301,8 +248,16 @@ class PlayerService
         if(count($team->getPlayers()) > 9) {
             $fakePlayer = $team->getFakePlayerSamePosition($realPlayer->getPosition())->first();
 
-            $this->entityManager->remove($fakePlayer);
-            $this->entityManager->flush();
+            if (!$fakePlayer instanceof Player) {
+                $fakePlayer = $team->getPlayers()->filter(function (Player $player) {
+                    return $player->getIsRealPlayer() === false;
+                })->first();
+            }
+
+            if ($fakePlayer instanceof Player) {
+                $this->entityManager->remove($fakePlayer);
+                $this->entityManager->flush();
+            }
         }
     }
 
@@ -336,7 +291,8 @@ class PlayerService
         ]);
 
         $playerAttribute->setValue(
-            ($playerAttribute->getValue() + $trainingCamp->getSkillPoints()) >=99 ? 99 : $playerAttribute->getValue() + $trainingCamp->getSkillPoints()
+            ($playerAttribute->getValue() + $trainingCamp->getSkillPoints()) >=99
+                ? 99 : $playerAttribute->getValue() + $trainingCamp->getSkillPoints()
         );
 
         $this->entityManager->persist($playerAttribute);
